Atualizando atvPaciente. Adicionei uma conexão binária entre as classes Medico e Paciente. Outra mudança foi a adição dos metodos Post, Put e Delete para ambas as Classes

+++ src/main/java/com/curso/repositories/MedicoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MedicoRepository extends JpaRepository<Medico, Long> {
+public interface MedicoRepository extends JpaRepository<Medico, Integer> {
 }
